requirements.txt
mss>=6.1.0
pyperclip>=1.8.2
requests>=2.28.0
pynput>=1.7.6
python-dotenv>=0.20.0
